requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
django-rest-knox==4.2.0
djangorestframework==3.14.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
openai==1.14.3
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
robotframework==7.0
robotframework-requests==0.9.6
rsa==4.9
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
